+++ Sources/StreamChat/Controllers/ChannelListController/ChannelListController.swift
@@ -154,12 +154,14 @@ public class ChatChannelListController: DataController, DelegateCallable, DataSt
     private func updateChannelList(
         _ completion: ((_ error: Error?) -> Void)? = nil
     ) {
+        let limit = query.pagination.pageSize
         worker.update(
             channelListQuery: query
         ) { result in
             switch result {
-            case .success:
+            case let .success(channels):
                 self.state = .remoteDataFetched
+                self.hasLoadedAllPreviousChannels = channels.count < limit
                 self.callback { completion?(nil) }
             case let .failure(error):
                 self.state = .remoteDataFetchFailed(ClientError(with: error))
@@ -317,6 +319,28 @@ public class ChatChannelListController: DataController, DelegateCallable, DataSt
                 self.callback { completion?(nil) }
             case let .failure(error):
                 self.callback { completion?(error) }
+            }
+        }
+    }
+
+    func resetQuery(
+        watchedAndSynchedChannelIds: Set<ChannelId>,
+        synchedChannelIds: Set<ChannelId>,
+        completion: @escaping (Result<(synchedAndWatched: [ChatChannel], unwanted: Set<ChannelId>), Error>) -> Void
+    ) {
+        let pageSize = query.pagination.pageSize
+        worker.resetChannelsQuery(
+            for: query,
+            pageSize: pageSize,
+            watchedAndSynchedChannelIds: watchedAndSynchedChannelIds,
+            synchedChannelIds: synchedChannelIds
+        ) { [weak self] result in
+            switch result {
+            case let .success((newChannels, unwantedCids)):
+                self?.hasLoadedAllPreviousChannels = newChannels.count < pageSize
+                completion(.success((newChannels, unwantedCids)))
+            case let .failure(error):
+                completion(.failure(error))
             }
         }
     }